requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

asgiref==3.7.2
attrs==23.1.0
Brotli==1.1.0
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
cloudinary==1.36.0
distlib==0.3.7
django==4.2.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
filelock==3.13.1
gunicorn==21.2.0
identify==2.5.32
idna==3.4
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
netifaces==0.11.0
nodeenv==1.8.0
packaging==23.2
platformdirs==3.11.0
pre-commit==3.5.0
psycopg2-binary==2.9.9
pydotplus==2.0.2
pyjwt==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.1
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.24.6
whitenoise==6.6.0
